+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.2.6
+- Added status codes for output generate from publish.
+
+## 0.2.5
+- Added domain option in CLI
+
+## 0.2.4
+- Removed unused libraries and updated a few library versions.
+
+## 0.2.3
+- Implemented array of Json objects as input and output in Json format in CLI
+- Added Debug flag
+
+## 0.2.2
+- Added eventId to output format in REMREM publish
+- Implemented get route key from messaging library.
+
 ## 0.2.1
 - Fix authentication required even when activedirectory.enabled=false